+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## `2026-02-28`
+
+- Added a `withpass` script to read passwords from standard input and pass them
+  to a command as an environment variable.
+  ([#687](https://github.com/ianlewis/dotfiles/issues/687).
+
 ## `2026-02-25`
 
 - Added `krew` as a global CLI tool.